+++ probeye/definition/noise_model.py
@@ -2,13 +2,12 @@
 import numpy as np
 
 # local imports
-from probeye.subroutines import len_or_one
-from probeye.subroutines import list2dict, make_list
+from probeye.subroutines import make_list, translate_prms_def
 
 
 class NoiseModelBase:
-    def __init__(self, dist, prms_def, sensors, name=None, correlation=None,
-                 noise_type='additive'):
+    def __init__(self, dist, prms_def, sensors, name=None, corr=None,
+                 corr_model=None, noise_type='additive'):
         """
         Parameters
         ----------
@@ -21,16 +20,18 @@ class NoiseModelBase:
             prms_def = ['mu', 'sigma'] means that the noise parameter vector
             has two elements, the first of which gives the value of 'mu' and the
             second gives the value of 'sigma'.
-        sensors : str, list[str]
-            Names of the sensors that are required to evaluate the noise model.
+        sensors : obj[Sensor], list[Sensor]
+            Sensor objects that are required to evaluate the noise model.
         name : str, None, optional
             Unique name of the noise model. This name is None, if the user does
             not specify it when adding the noise model to the problem. It is
             then named automatically before starting the inference engine.
-        correlation : dict, None, optional
+        corr : dict, None, optional
             Defines the correlation model. So far this is just a placeholder.
             It is not clear yet how exactly the correlation should be defined.
             When it is set to None, all sensors/sensor elements are independent.
+        corr_model : str, optional
+            Defines the correlation function to be used in case corr isn't None.
         noise_type : str, optional
             Either 'additive', 'multiplicative' or 'other'. Defines if the error
             is computed via [prediction - measurement] ('additive') or via
@@ -38,11 +39,12 @@ class NoiseModelBase:
             i.e., non-standard fashion.
         """
         self.dist = dist
-        self.prms_def = list2dict(make_list(prms_def))
-        self.prms_dim = len_or_one(make_list(prms_def))
+        self.prms_def, self.prms_dim = translate_prms_def(prms_def)
         self.sensors = make_list(sensors)
+        self.sensor_names = [sensor.name for sensor in self.sensors]
         self.name = name
-        self.correlation = correlation
+        self.corr = corr
+        self.corr_model = corr_model
         self.noise_type = noise_type
         # this is a list of experiment names, that relate to the noise model;
         # the list will be filled after experiments have been added to the
@@ -86,7 +88,7 @@ class NoiseModelBase:
             exp_dict = self.problem_experiments[exp_name]
             forward_models.add(exp_dict['forward_model'])
             sensor_names_exp = [*exp_dict['sensor_values'].keys()]
-            for sensor_name in self.sensors:
+            for sensor_name in self.sensor_names:
                 if sensor_name not in sensor_names_exp:
                     raise RuntimeError(
                         f"Experiment '{exp_name}' does not contain a sensor "
@@ -123,7 +125,7 @@ class NoiseModelBase:
             1D numpy arrays representing the model errors as values.
         """
         # prepare the dictionary keys
-        model_error_dict = {name: np.array([]) for name in self.sensors}
+        model_error_dict = {name: np.array([]) for name in self.sensor_names}
 
         # fill the dictionary with model error vectors
         for exp_name in self.experiment_names:
@@ -133,7 +135,7 @@ class NoiseModelBase:
             me_dict = self.error_function(ym_dict, ye_dict)
             model_error_dict =\
                 {name: np.append(model_error_dict[name], me_dict[name])
-                 for name in self.sensors}
+                 for name in self.sensor_names}
 
         return model_error_dict
 
@@ -175,7 +177,7 @@ class NoiseModelBase:
         """
         # for each sensor, its own error metric is used to compute the error
         error_dict = {name: ym_dict[name] - ye_dict[name]
-                      for name in self.sensors}
+                      for name in self.sensor_names}
         return error_dict
 
     def error_function_multiplicative(self, ym_dict, ye_dict):
@@ -197,7 +199,7 @@ class NoiseModelBase:
         """
         # for each sensor, its own error metric is used to compute the error
         error_dict = {name: ym_dict[name] / ye_dict[name] - 1.0
-                      for name in self.sensors}
+                      for name in self.sensor_names}
         return error_dict
 
     def error_function_other(self, ym_dict, ye_dict):
@@ -239,35 +241,15 @@ class NormalNoiseModel(NoiseModelBase):
     A general Gaussian (normal) noise model with or without correlations.
     """
 
-    def __init__(self, prms_def, sensors, name=None, correlation=None,
+    def __init__(self, prms_def, sensors, name=None, corr=None, corr_model=None,
                  noise_type='additive'):
         """
-        Parameters
-        ----------
-        prms_def : str, list[str], dict
-            The parameters of the noise model. Currently, this noise model needs
-            to have a 'std'-parameter which describes the standard deviation.
-            Also the parameter 'mean' can be given.
-        sensors : str, list[str]
-            Names of the sensors that are required to evaluate the noise model.
-        name : str, None, optional
-            Unique name of the noise model. This name is None, if the user does
-            not specify it when adding the noise model to the problem. It is
-            then named automatically before starting the inference engine.
-        correlation : dict, None, optional
-            Defines the correlation model. So far this is just a placeholder.
-            It is not clear yet how exactly the correlation should be defined.
-            When it is set to None, all sensors/sensor elements are independent.
-        noise_type : str, optional
-            Either 'additive', 'multiplicative' or 'other'. Defines if the error
-            is computed via [prediction - measurement] ('additive') or via
-            [prediction/measurement-1] ('multiplicative') or in some 'other'
-            i.e., non-standard fashion.
+        See docstring of NoiseModelBase for information on the input arguments.
         """
 
         # initialize the base class with given input
-        super().__init__('normal', prms_def, sensors, name=name,
-                         correlation=correlation, noise_type=noise_type)
+        super().__init__('normal', prms_def, sensors, name=name, corr=corr,
+                         corr_model=corr_model, noise_type=noise_type)
 
         # check that at the standard deviation is provided (this can be either
         # as a constant or a latent parameter, but it has to be given)