+++ probeye/definition/prior.py
@@ -1,5 +1,5 @@
 # local imports
-from probeye.subroutines import list2dict
+from probeye.subroutines import translate_prms_def
 
 
 class PriorBase:
@@ -39,8 +39,9 @@ class PriorBase:
         # then 'a' is the reference variable; note that the conversion to a
         # dictionary via list2dict is due to the possibility of using local
         # parameter names, which however is not intended to be used for priors
-        self.hyperparameters = list2dict(prms_def)
-        self.prms_def = list2dict([ref_prm] + prms_def)
+        self.hyperparameters, _ = translate_prms_def(prms_def)
+        self.prms_def, _ = translate_prms_def(
+            [ref_prm] + [*self.hyperparameters.values()])
 
     def __str__(self):
         """