+++ probeye/inference/emcee_/__init__.py
@@ -1,4 +1,5 @@
 # module imports
+from probeye.inference.emcee_ import correlation_models
 from probeye.inference.emcee_ import noise_models
 from probeye.inference.emcee_ import priors
 from probeye.inference.emcee_ import solver