+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # probeye changelog
 
-## 1.0.4 (2021-Oct-05)
+## 1.0.4 (2021-Oct-??)
 ### Added
 - torch submodule (in inference submodule)
 - arviz-based post-processing for sampling methods (in postprocessing submodule)
+- integration test with a correlation example
+- new subroutines: flatten_generator, flatten, process_spatial_coordinates, translate_prms_def
+- new script: correlation_models.py in inference/emcee module 
 ### Changed
 - tool-independent noise model (in definition submodule)
 - tool-independent normal noise model (in definition submodule)
+- normal noise model for emcee (revised to be able to account for correlation)
 
 ## 1.0.3 (2021-Oct-04)
 ### Changed